Develop

+++ src/main/java/ru/rerumu/lists/views/ReadingRecordUpdateView.java
@@ -3,7 +3,7 @@ package ru.rerumu.lists.views;
 import java.time.LocalDateTime;
 
 public record ReadingRecordUpdateView(
-        Long statusId,
+        Integer statusId,
         LocalDateTime startDate,
         LocalDateTime endDate
 ) {
